+++ src/main/java/net/countercraft/movecraft/warfare/MovecraftWarfare.java
@@ -2,9 +2,9 @@ package net.countercraft.movecraft.warfare;
 
 import net.countercraft.movecraft.repair.MovecraftRepair;
 import net.countercraft.movecraft.util.Tags;
-import net.countercraft.movecraft.warfare.listener.BlockListener;
 import net.countercraft.movecraft.warfare.config.Config;
 import net.countercraft.movecraft.warfare.features.assault.AssaultManager;
+import net.countercraft.movecraft.warfare.features.assault.BlockListener;
 import net.countercraft.movecraft.warfare.features.assault.RegionDamagedSign;
 import net.countercraft.movecraft.warfare.features.assault.commands.AssaultCommand;
 import net.countercraft.movecraft.warfare.features.assault.commands.AssaultInfoCommand;
@@ -12,7 +12,6 @@ import net.countercraft.movecraft.warfare.features.assault.commands.AssaultRepai
 import net.countercraft.movecraft.warfare.features.siege.SiegeCommand;
 import net.countercraft.movecraft.warfare.features.siege.SiegeManager;
 import net.countercraft.movecraft.warfare.localisation.I18nSupport;
-import net.countercraft.movecraft.warfare.utils.WarfareRepair;
 import net.countercraft.movecraft.worldguard.MovecraftWorldGuard;
 import org.bukkit.Material;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -87,8 +86,6 @@ public final class MovecraftWarfare extends JavaPlugin {
 
             getServer().getPluginManager().registerEvents(new BlockListener(), this);
             getServer().getPluginManager().registerEvents(new RegionDamagedSign(), this);
-
-            new WarfareRepair(this);
         }
 
         getCommand("assaultinfo").setExecutor(new AssaultInfoCommand());
